requirements.txt
spotipy
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client